requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
